% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/setup.R
\name{get.access.key}
\alias{get.access.key}
\title{Get your BioGRID access key}
\usage{
get.access.key()
}
\value{
The users BioGRID access key
}
\description{
Retrieve the current access key. If it has not been set, the access
key is \code{NULL}
}
\keyword{authentication}

