% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/query.R
\name{get.interactions}
\alias{get.interactions}
\title{Retrieve interactions for one or more genes.}
\usage{
get.interactions(geneList, searchNames = TRUE, searchSynonyms = TRUE,
  searchIds = FALSE, searchBiogridIds = FALSE, taxId = 9606,
  selfInteractionsExcluded = TRUE, includeInteractorInteractions = FALSE,
  includeEvidence = TRUE, verbose = FALSE)
}
\arguments{
\item{geneList}{A character string with pipe delimited series of gene symbols}

\item{searchNames}{Use identifiers in \code{geneList} and search gene names}

\item{searchSynonyms}{Use identifiers in \code{geneList} and search gene synonyms}

\item{searchIds}{if \code{TRUE}, identifiers in \code{geneList} are searched for in Entrez Gene ID's.
For this to be used \code{searchNames} and \code{searchSynonyms} should be \code{FALSE}}

\item{searchBiogridIds}{if \code{TRUE}, identifiers in \code{geneList} are searched for in BioGRID ID's
For this to be used \code{searchNames}, \code{searchSynonyms} and \code{searchIds}
should be \code{FALSE}}

\item{taxId}{The species to consider. By default Homo Sapiens (9606)}

\item{selfInteractionsExcluded}{If \code{TRUE} ignore self interactions}

\item{includeInteractorInteractions}{If \code{TRUE}, then interactions between the interactors of the query targets are included in the return value}

\item{verbose}{If \code{TRUE} intermediate output is printed such as URLs and interaction counts}
}
\value{
A \code{data.frame} with interaction information
}
\description{
By default this function returns interactions for one or more gene
symbols, separated by the pipe ('|') symbol and considers Homo Sapiens
as the species. See \url{http://wiki.thebiogrid.org/doku.php/biogridrest#list_of_parameters}
for more details. This method caches queries to the BioGRID REST API, so that queries
for the same set of targets will be retrieved from cache. Note that cache is restricted
to the current session only.
}
\seealso{
\code{\link{get.organisms}}, \code{\link{set.access.key}}
}
\keyword{database}

