% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/query.R
\name{get.organisms}
\alias{get.organisms}
\title{Retrieve table of organisms and their taxonomy ID's}
\usage{
get.organisms()
}
\value{
A \code{data.frame} with two columns called "taxid" and "organism"
}
\description{
BioGRID supports protein protein interactions in multiple species.
By default queries consider all species, but can be restricted to
consider a specific species. This methods returns all species names
and their corresponding taxonomy ID
}
\seealso{
\code{link{set.access.key}}
}
\keyword{database}

