% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/setup.R
\name{set.access.key}
\alias{set.access.key}
\title{Set your BioGRID access key to be used for subsequent queries}
\usage{
set.access.key(key)
}
\arguments{
\item{key}{The users BioGRID access key}
}
\description{
To query for interactions from BioGRID requires that each request
include an access key. You should get this when signing up. Set the
access key for this session using this function.
}
\keyword{authentication}

